backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib
sqlalchemy
pydantic
mistralai
pypdf
python-dotenv
docling
